add legacy mania combo counter
