+++ osu.Game.Rulesets.Mania/ManiaSkinComponentLookup.cs
@@ -20,6 +20,7 @@ namespace osu.Game.Rulesets.Mania
     public enum ManiaSkinComponents
     {
         ColumnBackground,
+        ComboCounter,
         HitTarget,
         KeyArea,
         Note,
