+++ osu.Game/Skinning/LegacyManiaSkinConfiguration.cs
@@ -39,6 +39,7 @@ namespace osu.Game.Skinning
 
         public float HitPosition = DEFAULT_HIT_POSITION;
         public float LightPosition = (480 - 413) * POSITION_SCALE_FACTOR;
+        public float ComboPosition = 300 * POSITION_SCALE_FACTOR;
         public float ScorePosition = 300 * POSITION_SCALE_FACTOR;
         public bool ShowJudgementLine = true;
         public bool KeysUnderNotes;
