+++ osu.Game/Skinning/LegacySkinTransformer.cs
@@ -1,7 +1,10 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Linq;
 using osu.Framework.Audio.Sample;
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Containers;
 using osu.Game.Audio;
 using osu.Game.Rulesets.Objects.Legacy;
 using static osu.Game.Skinning.SkinConfiguration;
@@ -33,6 +36,30 @@ namespace osu.Game.Skinning
                 return new SampleVirtual();
 
             return base.GetSample(sampleInfo);
+        }
+
+        protected Drawable? ReplaceBaseComboCounter(ISkinComponentLookup lookup)
+        {
+            if (lookup is SkinComponentsContainerLookup containerLookup)
+            {
+                switch (containerLookup.Target)
+                {
+                    case SkinComponentsContainerLookup.TargetArea.MainHUDComponents:
+                        var components = base.GetDrawableComponent(lookup) as Container;
+
+                        if (components != null)
+                        {
+                            // some rulesets may provide its own combo counter; hide the default.
+                            // todo: this should be done in an elegant way per ruleset, defining which HUD skin components should be displayed.
+                            foreach (var legacyBaseComboCounter in components.OfType<LegacyBaseComboCounter>())
+                                legacyBaseComboCounter.HiddenByRulesetImplementation = false;
+                        }
+
+                        return components;
+                }
+            }
+
+            return null;
         }
     }
 }
